Added additional loaders in the summary for data sync and fixed loader icon issue + setup campaign file clean up  (#1744)
* code cleanup of setup campaign

* updated the draft utils

* taken latest pull

* Added summary loaders
