+++ health/micro-ui/web/micro-ui-internals/example/package.json
@@ -14,7 +14,7 @@
     "@egovernments/digit-ui-components": "0.0.2-beta.49",
     "@egovernments/digit-ui-module-core": "1.8.2-beta.13",
     "@egovernments/digit-ui-module-utilities": "1.0.1-beta.30",
-    "@egovernments/digit-ui-react-components": "1.8.2-beta.16",
+    "@egovernments/digit-ui-react-components": "1.8.2-beta.17",
     "@egovernments/digit-ui-module-hcmworkbench":"0.0.44",
     "@egovernments/digit-ui-module-campaign-manager": "0.2.1",
     "@egovernments/digit-ui-module-microplan":"0.0.1",
