+++ health/micro-ui/web/micro-ui-internals/example/public/index.html
@@ -10,11 +10,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1" />
   <meta name="theme-color" content="#00bcd1" />
   <title>DIGIT</title>
-  <!-- <link
-    rel="stylesheet"
-    href="https://unpkg.com/@egovernments/digit-ui-css@1.8.0-alpha.6/dist/index.css"
-  /> -->
-  <link rel="stylesheet" href="https://unpkg.com/@egovernments/digit-ui-css@1.8.2-beta.34/dist/index.css" />
+  <link rel="stylesheet" href="https://unpkg.com/@egovernments/digit-ui-css@1.8.2-beta.49/dist/index.css" />
   <link rel="stylesheet" href="https://unpkg.com/@egovernments/digit-ui-components-css@0.0.2-beta.39/dist/index.css" />
   <link rel="stylesheet" href="https://unpkg.com/@egovernments/digit-ui-health-css@0.1.19/dist/index.css" />
