+++ health/micro-ui/web/micro-ui-internals/package.json
@@ -49,7 +49,7 @@
     "*.{js,css,md}": "prettier --write"
   },
   "dependencies": {
-    "@egovernments/digit-ui-react-components": "1.8.2-beta.16",
+    "@egovernments/digit-ui-react-components": "1.8.2-beta.17",
     "@egovernments/digit-ui-components": "0.0.2-beta.49",
     "ajv": "8.12.0",
     "lodash": "4.17.21",
