+++ health/micro-ui/web/micro-ui-internals/packages/css/src/pages/employee/campaign.scss
@@ -131,7 +131,7 @@
   height: 31.25rem;
 }
 .stepper-card{
-  height: 36.5rem;
+  height: 30rem;
   margin-bottom: 1.5rem !important;
 }
 .stepper-subheader{
