+++ health/micro-ui/web/micro-ui-internals/packages/modules/campaign-manager/package.json
@@ -20,7 +20,7 @@
   "dependencies": {
     "@cyntler/react-doc-viewer": "1.10.3",
     "@egovernments/digit-ui-components": "0.0.2-beta.49",
-    "@egovernments/digit-ui-react-components": "1.8.2-beta.16",
+    "@egovernments/digit-ui-react-components": "1.8.2-beta.17",
     "@rjsf/core": "5.10.0",
     "@rjsf/utils": "5.10.0",
     "@rjsf/validator-ajv8": "5.10.0",
