+++ health/micro-ui/web/micro-ui-internals/packages/modules/campaign-manager/src/components/DeliveryDetailsSummary.js
@@ -1,7 +1,7 @@
-import React, { Fragment, useEffect, useRef, useState } from "react";
+import React, { Fragment, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { useHistory } from "react-router-dom";
-import {  EditIcon, Header, Loader, ViewComposer } from "@egovernments/digit-ui-react-components";
+import {  EditIcon, Header, Loader, LoaderWithGap, ViewComposer } from "@egovernments/digit-ui-react-components";
 import {  Toast, Card, Stepper, TextBlock } from "@egovernments/digit-ui-components";
 
 
@@ -223,7 +223,7 @@ const DeliveryDetailsSummary = (props) => {
     }
   }, [props?.props?.summaryErrors]);
 
-  const { isLoading, data, error, refetch } = Digit.Hooks.campaign.useSearchCampaign({
+  const { isLoading, data, error, refetch,isFetching } = Digit.Hooks.campaign.useSearchCampaign({
     tenantId: tenantId,
     filter: {
       ids: [id],
@@ -300,9 +300,7 @@ const DeliveryDetailsSummary = (props) => {
     },
   });
 
-  if (isLoading) {
-    return <Loader />;
-  }
+
   const closeToast = () => {
     setShowToast(null);
   };
@@ -331,9 +329,13 @@ const DeliveryDetailsSummary = (props) => {
     } else if (currentStep === 2) setKey(9);
     else setKey(8);
   };
+  if (isLoading) {
+    return <Loader />;
+  }
 
   return (
     <>
+      {(isLoading || (!data && !error) || isFetching) &&<LoaderWithGap text={t("DATA_SYNC_WITH_SERVER")} />}
       <div className="container-full">
         <div className="card-container">
           <Card className="card-header-timeline">
