+++ health/micro-ui/web/package.json
@@ -18,7 +18,7 @@
     "@egovernments/digit-ui-module-workbench": "1.0.2-beta.7",
     "@egovernments/digit-ui-module-core": "1.8.2-beta.13",
     "@egovernments/digit-ui-module-hrms": "1.8.0-beta.2",
-    "@egovernments/digit-ui-react-components": "1.8.2-beta.16",
+    "@egovernments/digit-ui-react-components": "1.8.2-beta.17",
     "@egovernments/digit-ui-components": "0.0.2-beta.49",
     "@egovernments/digit-ui-module-dss": "1.8.0-beta",
     "@egovernments/digit-ui-module-common": "1.8.0-beta",
