+++ health/micro-ui/web/workbench/inter-package.json
@@ -46,7 +46,7 @@
   "dependencies": {
     "lodash": "4.17.21",
     "microbundle-crl": "0.13.11",
-    "@egovernments/digit-ui-react-components": "1.8.2-beta.1",
+    "@egovernments/digit-ui-react-components": "1.8.2-beta.17",
     "@egovernments/digit-ui-components": "0.0.2-beta.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
