+++ health/micro-ui/web/workbench/package.json
@@ -17,7 +17,7 @@
     "@egovernments/digit-ui-components": "0.0.2-beta.49",
     "@egovernments/digit-ui-module-core": "1.8.2-beta.25",
     "@egovernments/digit-ui-module-utilities": "1.0.1-beta.35",
-    "@egovernments/digit-ui-react-components": "1.8.2-beta.16",
+    "@egovernments/digit-ui-react-components": "1.8.2-beta.17",
     "@egovernments/digit-ui-module-hcmworkbench":"0.0.44",
     "@egovernments/digit-ui-module-campaign-manager": "0.2.1",
     "babel-loader": "8.1.0",
